Make the to state optional
Since we cannot offer pattern matching on the struct as of now, we
cannot inject multiple clauses with the same name.
This is desirable in many cases, and as a workaround, to to state can be
skipped.
It is assumed that the transition function will change the state on it's
own.
`:persist` optionstill works as it did.

Adds a state machine with such an event